Feat: Moved Relevant Components to Dashboard components folder
