+++ frontend/src/app/dashboard/page.tsx
@@ -1,11 +1,8 @@
 "use server";
 
-import DashboardClient from "@/components/custom/DashboardClient";
 import { createEmbedding } from "../actions/createEmbedding";
+import DashboardClient from "./components/DashboardClient";
 
 export default async function DasbhoardPage() {
-  return (
-
-    <DashboardClient createEmbedding={createEmbedding} />
-  );
+  return <DashboardClient createEmbedding={createEmbedding} />;
 }
